% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharToNum.R
\name{CharToNum}
\alias{CharToNum}
\title{CharToNum}
\usage{
CharToNum(data, vars, rename = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{vars}{a vector of variable names, which are character variables in the dataframe need to convert to numeric type}

\item{rename}{Boolean value, whether to rename the variable name and create a new variable (add ".numeric' suffix), with default FALSE}
}
\value{
NumToChar returns a dataframe.
}
\description{
The function CharToNum is used to convert character variables to numeric type variables.
}
\examples{
str(NumToChar(iris, 'Sepal.Length'))
str(NumToChar(iris, c('Sepal.Length', 'Sepal.Width'), rename = TRUE))
}
