% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumToChar.R
\name{NumToChar}
\alias{NumToChar}
\title{NumToChar}
\usage{
NumToChar(data, vars, rename = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{vars}{a vector of variable names, which are numeric variables in the dataframe need to convert to character type}

\item{rename}{Boolean value, whether to rename the variable name and create a new variable (add ".char' suffix), with default FALSE}
}
\value{
NumToChar returns a dataframe.
}
\description{
The function NumToChar is used to convert numeric variables to character type variables.
}
\examples{
str(NumToChar(iris, 'Sepal.Length'))
str(NumToChar(iris, c('Sepal.Length', 'Sepal.Width'), rename = TRUE))
}
