% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegVarImp.R
\name{RegVarImp}
\alias{RegVarImp}
\title{RegVarImp}
\usage{
RegVarImp(data, yvar, xvars = "all", method = c("Lm", "SRTree", "MRTree"),
  rtreecontrol = rpart.control(minsplit = 2, cp = 1e-04, xval = 10, maxdepth =
  30), round.num = 5)
}
\arguments{
\item{data}{a dataframe}

\item{round.num}{exact digits}

\item{var}{a variable name}

\item{pdf}{Boolean value, whether to save as a pdf file, with default FALSE}

\item{savedir}{the pdf file save directory, like "D:/test/"}
}
\description{
The function RegVarImp.
}
