% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleVarPlot.R
\name{SingleVarPlot}
\alias{SingleVarPlot}
\title{SingleVarPlot}
\usage{
SingleVarPlot(data, var, round.num = 3, pdf = FALSE, savedir)
}
\arguments{
\item{data}{a dataframe}

\item{var}{a variable name}

\item{round.num}{exact digits}

\item{pdf}{Boolean value, whether to save as a pdf file, with default FALSE}

\item{savedir}{the pdf file save directory, like "D:/test/"}
}
\value{
VarGroup returns a dataframe.
}
\description{
The function SingleVarPlot is used to group the vector.
}
\examples{
SingleVarPlot(data = iris, var = "Species")
SingleVarPlot(data = iris, var = "Sepal.Length", round.num = 2)
SingleVarPlot(data = iris, var = "Sepal.Length", round.num = 2, pdf = TRUE, savedir = "D:/")
}
\seealso{
VarGroupText
}
