% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToFactor.R
\name{ToFactor}
\alias{ToFactor}
\title{ToFactor}
\usage{
ToFactor(data, vars, rename = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{vars}{a vector of variable names, which need to convert to factor type}

\item{rename}{Boolean value, whether to rename the variable name and create a new variable (add ".factor' suffix), with default FALSE}
}
\value{
NumToChar returns a dataframe.
}
\description{
The function ToFactor is used to convert variables to factor type variables.
}
\examples{
str(NumToChar(iris, 'Sepal.Length'))
str(NumToChar(iris, c('Sepal.Length', 'Sepal.Width'), rename = TRUE))
}
