% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueReplace.R
\name{ValueReplace}
\alias{ValueReplace}
\title{ValueReplace}
\usage{
ValueReplace(data, vars, oris, new, rename = FALSE)
}
\arguments{
\item{data}{a dataframe}

\item{vars}{a vector of variable names, which need to replace value}

\item{oris}{a vector of original values}

\item{new}{a new value}

\item{rename}{Boolean value, whether to rename the variable name and create a new variable, with default FALSE}
}
\value{
ValueReplace returns a dataframe.
}
\description{
The function ValueReplace is used to replace the variable value.
}
\examples{
ValueReplace(iris, 'Sepal.Length', ori = 2, 1 )
ValueReplace(iris, 'Sepal.Length', ori = c(1.7,5.1), 1,TRUE)
}
