% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarDel.R
\name{VarDel}
\alias{VarDel}
\title{VarDel}
\usage{
VarDel(data, vars)
}
\arguments{
\item{data}{a dataframe}

\item{vars}{a vector of variable names, which need to be deleted}
}
\value{
VarDel returns a dataframe.
}
\description{
The function VarDel is used to delete the variables in the dataframe.
}
\examples{
VarDel(iris, "Species")
}
