% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGroup.R
\name{VarGroup}
\alias{VarGroup}
\title{VarGroup}
\usage{
VarGroup(vector, gcondition, gname)
}
\arguments{
\item{vector}{a vector}

\item{gcondition}{a vector of group conditions, matches gname}

\item{gname}{a vector of group names, matches gcondition}
}
\value{
VarGroup returns a dataframe.
}
\description{
The function VarGroup is used to group the vector.
}
\examples{
#' #(1)factor
basedata = VarGroupText(iris$Sepal.Length, "bin", bins = 10, nametype = "factor")
VarGroup(iris$Sepal.Length, basedata$condition, basedata$groupname)

#(2)numeric
basedata = VarGroupText(iris$Sepal.Length, "bin", bins = 10, nametype = "numeric", valuetype = "midpoint")
VarGroup(iris$Sepal.Length, basedata$condition, basedata$groupname)
}
\seealso{
VarGroupText
}
