% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarGroupText.R
\name{VarGroupText}
\alias{VarGroupText}
\title{VarGroupText}
\usage{
VarGroupText(vector, method = c("bin", "quantile"), bins = 10,
  nametype = c("factor", "numeric"), valuetype = c("midpoint", "lower",
  "upper"))
}
\arguments{
\item{vector}{a vector}

\item{method}{"bin" or "quantile"}

\item{bins}{bins}

\item{nametype}{the data type of group name, "factor" or "numeric"}

\item{valuetype}{group representation value, "midpoint" or "lower" or "upper"}
}
\value{
VarGroupText returns a dataframe.
}
\description{
The function VarGroupText is used to group the numeric vector.
}
\examples{
VarGroupText(iris$Sepal.Length, "bin", bins = 10, nametype = "factor")
VarGroupText(iris$Sepal.Length, "bin", bins = 10, nametype = "numeric", valuetype = "midpoint")
}
