% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarStat.R
\name{VarStat}
\alias{VarStat}
\title{VarStat}
\usage{
VarStat(data, var_type, nstat = c("Miss", "MissRate", "N", "Mean", "Sd", "Cv",
  "Skewness", "Kurtosis", "Min", "Max", "Median", "Num", "Outlier"),
  cstat = c("Miss", "MissRate", "N", "Num", "Mode"), round.num = 2)
}
\arguments{
\item{data}{a dataframe}

\item{var_type}{a result returned by the VarType function}

\item{nstat}{a vector of the descriptive statistics of numerical variables, you can choose c("Miss", "MissRate", "N", "Mean", "Sd", "Cv", "Skewness", "Kurtosis","Min", "Max", "Median", "Num", "Outlier") one or more}

\item{cstat}{a vector of the descriptive statistics of non-numerical variables, you can choose c("Miss", "MissRate", "N", "Num", "Mode") one or more}

\item{round.num}{exact digits}
}
\value{
VarStat returns two dataframe.
}
\description{
The function VarStat is used to calculate the variable description statistics.
}
\examples{
VarStat(iris,VarType(iris))
}
\seealso{
VarStat
}
