% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarType.R
\name{VarType}
\alias{VarType}
\title{VarType}
\usage{
VarType(data, dtypes = c("integer", "numeric", "character", "factor"),
  threshold = 50, showtxt = FALSE)
}
\arguments{
\item{data}{a dataframe to be determined}

\item{dtypes}{candidate data type, with default parameter c("integer", "numeric", "character", "factor")}

\item{threshold}{a threshold that determines whether a variable may be a discrete or category variable, with default 50}

\item{showtxt}{Boolean value, whether to output text conclusions, with default FALSE}

\item{dtype}{Boolean value, determination result of whether the variable belongs to each data type of the "dtypes" parameter}

\item{value.num}{the number of different values of the variable}

\item{value.num.prop}{the ratio of the number of different values of the variable to the total number of samples}

\item{note}{note}

\item{type}{variable data type summary results}
}
\value{
VarType returns a dataframe of the data type determination result.

An object of class "VarType" is a dataframe containing at least the following components:
}
\description{
The function VarType is used to determine the data type of each variable in the dataframe.
}
\examples{
VarType(iris)
}
